requirements.txt
torch
numpy
Pillow
faiss-cpu
matplotlib
open-clip-torch
fastapi
uvicorn
python-dotenv
pymongo
requests
python-multipart